Introduce more OOP design to the program
The program tries to do everything in a few classes, particularly the
Duke class.

It is harder to modify the code as it all happens in one class.

Adding new classes will make modifying the code in the future easier.
