+++ src/main/java/Main.java
@@ -1,0 +1,7 @@
+import utility.Duke;
+
+public class Main {
+    public static void main(String[] args) {
+        new Duke().run();
+    }
+}
