+++ src/main/java/tasks/Deadline.java
@@ -27,6 +27,7 @@ public class Deadline extends Task{
     public Deadline(String description, String by) {
         super(description);
         String[] dateAndTime = by.trim().split(" ", 2);
+
         if (Pattern.matches(YMD_REGEX, dateAndTime[0])) {
             try {
                 DateFormat format1 = new SimpleDateFormat(YMD_DATE_FORMAT);
